Replace mypipe.py with intended new version
